rename(Fallback): common 폴더로 이동
